refactoring of types
